refactor(cli): share the multi-path rule loader across commands
`backend convert` owned a private `load_collection_multi` that walks one or
more file/directory paths into a single `SigmaCollection`. Promote it to a
crate-level helper next to the existing single-path `load_collection` so the
upcoming `rule coverage` command consumes the same loader and the two cannot
drift on how repeated `--rules` paths are merged or how load errors exit.

Behavior-neutral: the function body is unchanged and `backend convert` now
calls `crate::load_collection_multi`.
